Lines up to tape now!
