+++ src/main/java/frc/robot/commands/autoCommands/AutoAlignHorizontalCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.complexCommands;
+package frc.robot.commands.autoCommands;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.util.Units;
@@ -6,28 +6,26 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.swerve.SwerveDrive;
 import frc.robot.subsystems.vision.Vision;
 
-public class AutoAlignRotationalCommand extends CommandBase {
+public class AutoAlignHorizontalCommand extends CommandBase {
 
 	private int limelightId;
 	private SwerveDrive swerve;
 	private Vision vision;
+	private double timeThreshold;
+	private double translationThreshold;
+	private int timeCorrect;
 	private PIDController pid;
-	private int timeThreshold;
-	private double rotationalThreshold;
-	private int timesCorrect;
 
-	public AutoAlignRotationalCommand(
-		int limelightId,
-		int timeThreshold,
-		double rotationThreshold
+	public AutoAlignHorizontalCommand(
+		int limelightId
 	) {
 		this.swerve = SwerveDrive.getInstance();
 		this.vision = Vision.getInstance();
 		this.limelightId = limelightId;
+		timeThreshold = 0.5;
+		this.timeCorrect = 0;
+		translationThreshold = 1;
 		this.pid = new PIDController(1, 0, 0);
-		this.timeThreshold = timeThreshold;
-		this.rotationalThreshold = rotationThreshold;
-		this.timesCorrect = 0;
 		addRequirements(swerve, vision);
 		pid.reset();
 		pid.setSetpoint(0);
@@ -40,19 +38,19 @@ public class AutoAlignRotationalCommand extends CommandBase {
 		);
 		swerve.driveRobotCentric(
 			0,
-			0,
-			pid.calculate(horizontalAngleOffset) / 10
+			pid.calculate(horizontalAngleOffset) / 10,
+			0
 		);
-		if (Math.abs(horizontalAngleOffset) < rotationalThreshold) {
-			timesCorrect++;
+		if (Math.abs(horizontalAngleOffset) < translationThreshold) {
+			timeCorrect++;
 		} else {
-			timesCorrect = 0;
+			timeCorrect = 0;
 		}
 	}
 
 	@Override
 	public boolean isFinished() {
-		return timesCorrect >= timeThreshold * 50;
+		return timeCorrect >= timeThreshold * 50;
 	}
 
 	@Override
